+++ src/main/java/hibernate/ddl/DropQueryBuilder.java
@@ -1,6 +1,6 @@
 package hibernate.ddl;
 
-import hibernate.entity.EntityClass;
+import hibernate.entity.meta.EntityClass;
 
 public class DropQueryBuilder {
