+++ src/main/java/hibernate/dml/DeleteQueryBuilder.java
@@ -1,6 +1,6 @@
 package hibernate.dml;
 
-import hibernate.entity.column.EntityColumn;
+import hibernate.entity.meta.column.EntityColumn;
 
 public class DeleteQueryBuilder {
