+++ src/main/java/hibernate/dml/SelectQueryBuilder.java
@@ -1,6 +1,6 @@
 package hibernate.dml;
 
-import hibernate.entity.column.EntityColumn;
+import hibernate.entity.meta.column.EntityColumn;
 
 import java.util.List;
