+++ src/main/java/hibernate/entity/meta/column/EntityId.java
@@ -1,4 +1,4 @@
-package hibernate.entity.column;
+package hibernate.entity.meta.column;
 
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
