+++ src/test/java/hibernate/ddl/strategy/IdIdentityOptionGenerateStrategyTest.java
@@ -1,6 +1,6 @@
 package hibernate.ddl.strategy;
 
-import hibernate.entity.column.EntityId;
+import hibernate.entity.meta.column.EntityId;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
