+++ src/test/java/hibernate/ddl/strategy/PrimaryKetOptionGenerateStrategyTest.java
@@ -1,7 +1,7 @@
 package hibernate.ddl.strategy;
 
-import hibernate.entity.column.EntityField;
-import hibernate.entity.column.EntityId;
+import hibernate.entity.meta.column.EntityField;
+import hibernate.entity.meta.column.EntityId;
 import jakarta.persistence.Id;
 import org.junit.jupiter.api.Test;
