+++ src/test/java/hibernate/dml/UpdateQueryBuilderTest.java
@@ -1,7 +1,7 @@
 package hibernate.dml;
 
-import hibernate.entity.column.EntityColumn;
-import hibernate.entity.column.EntityField;
+import hibernate.entity.meta.column.EntityColumn;
+import hibernate.entity.meta.column.EntityField;
 import jakarta.persistence.*;
 import org.junit.jupiter.api.Test;
