+++ src/test/java/hibernate/entity/EntityManagerImplTest.java
@@ -3,6 +3,10 @@ package hibernate.entity;
 import database.DatabaseServer;
 import database.H2;
 import hibernate.ddl.CreateQueryBuilder;
+import hibernate.entity.meta.EntityClass;
+import hibernate.entity.persistencecontext.EntityKey;
+import hibernate.entity.persistencecontext.EntitySnapshot;
+import hibernate.entity.persistencecontext.SimplePersistenceContext;
 import jakarta.persistence.*;
 import jdbc.JdbcTemplate;
 import jdbc.ReflectionRowMapper;
