+++ src/test/java/hibernate/entity/meta/EntityTableNameTest.java
@@ -1,4 +1,4 @@
-package hibernate.entity;
+package hibernate.entity.meta;
 
 import jakarta.persistence.Table;
 import org.junit.jupiter.api.Test;
