+++ src/test/java/hibernate/entity/meta/column/ColumnTypeTest.java
@@ -1,5 +1,6 @@
-package hibernate.entity.column;
+package hibernate.entity.meta.column;
 
+import hibernate.entity.meta.column.ColumnType;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
