+++ src/test/java/hibernate/entity/meta/column/EntityColumnsTest.java
@@ -1,5 +1,7 @@
-package hibernate.entity.column;
+package hibernate.entity.meta.column;
 
+import hibernate.entity.meta.column.EntityColumn;
+import hibernate.entity.meta.column.EntityColumns;
 import jakarta.persistence.Column;
 import jakarta.persistence.Id;
 import jakarta.persistence.Transient;
